Added new ammunition textures and models for all ammunition Added landmines Fixed and improved nuke explosion, now emits light shafts, creates a wasteland biome around it and destroys certain blocks (trees, plants, circuits) in medium radius and leaves in a longer radius, nukes create fire blocks and apply Radiation and Nuclear Heat potion effects to players nearby Added Radio Explosives Added upgrades for Light Engineer Armor, WIP Field howitzer can now be pushed by players (operators "riding" it) Added replacement textures for IE bullets to match II's caliber based textures Fixed Artillery Howitzer rendering Fixed Ballistic Computer rendering Reworked white phosphorus bullet component Added missing textures for carbon dioxide Added GUIs and functionality for the emplacements, they can now target entities using filters, receive and send data commands, reload and repair, now they also require ammo and energy (usage depends on weapon installed, see config) Added new sounds for the emplacements Added new sounds for bullet hits depending on material and a wind sound for bullet fly-by Fixed bullets/blocks using PenetrationHandlerGeneral as default and not their type specific one Added new explosion sounds Added new machine sounds Added lots of new methods to String circuit Added new circuits: Document Circuit (for reading books and other text containing items), Type Conversion Circuit (for checking nulls, converting types, i.e. int to boolean or string), Entity Circuit (for handling entities detected by emplacements) Added duraluminium, the lightweight aircraft alloy (no use Yet™) Added Immersive Energy compat for shrapnel Partially added Immersive Posts compat (needs ASM works or an API from IPo side) Fixed naval mines Fixed the NBT Multiblock system for conveyor checking Reworked bullet items, now casings are separate and only a crafting ingredient Finished the Filler (before called the Casing Filler) multiblock, now it's not limited to ammo and uses new DustStack API (dusts stored in similar way to fluids) Tweaks to bullet values, now velocity is dependent on type and not on the gun, bullets cores are locked for specific ammo types "New" GUIs for data machines (smoothened the background) Fixed gas masks not protecting from gasses Added rods for some II materials Fixed alarm siren Added the radar multiblock (WIP) Added machinegun and tesla coil as emplacement weapons Improved API, added comments Added the Mortar, a portable short range artillery weapon Upgraded Hans limbs rendering, Hanses can now kneel Added new sounds for trying to start motorbike with no fuel Added new Concealed (invisibility, but also hides armor) and Exposed (similar to glowing) potion effects Fixes to data machines Lots of renames and refactors, removed unnecessary comments and console spam Added Tripwire Coil for Tripmines Fixes for TileEntities' blockbounds, now you can place blocks like levers on most of them Lots of miscellaneous fixes missed in this changelog Optimized textures

+++ src/main/java/pl/pabilo8/immersiveintelligence/client/render/multiblock/metal/FlagpoleRenderer.java
@@ -58,7 +58,7 @@ public class FlagpoleRenderer extends TileEntitySpecialRenderer<TileEntityFlagpo
 			if(!te.flag.isEmpty())
 			{
 				banner.setItemValues(te.flag, false);
-				float f = ((Math.abs((((getWorld().getTotalWorldTime()+partialTicks)%200)/200f)-0.5f)/0.5f)-0.5f)/0.5f;
+				double f = ((Math.abs((((getWorld().getTotalWorldTime()+partialTicks)%200)/200f)-0.5f)/0.5f)-0.5f)/0.5f;
 				ResourceLocation res = BannerTextures.BANNER_DESIGNS.getResourceLocation(banner.getPatternResourceLocation(), banner.getPatternList(), banner.getColorList());
 				if(res!=null)
 				{
@@ -93,13 +93,13 @@ public class FlagpoleRenderer extends TileEntitySpecialRenderer<TileEntityFlagpo
 		}
 	}
 
-	private void drawFlag(int n, float rot)
+	private void drawFlag(int n, double rot)
 	{
-		// TODO: 04.03.2021 make it work like it should
+		// TODO: 27.07.2021 fix uv
 		float length=40/(float)n;
 		for(int i = 0; i < n; i++)
 		{
-			GlStateManager.rotate(rot*6.5f*n*(n%2==0?1:-1),1,0,0);
+			GlStateManager.rotate((float)(rot*6.5f*n*(n%2==0?1:-1)),1,0,0);
 			ClientUtils.drawTexturedRect(0f,0f,1f,length/21f, 0.015625f,21/64f,0.015625f+(i*length/64f),(length*(i+1))/64f);
 			GlStateManager.translate(0f,0f,0.0625f);
 			ClientUtils.drawTexturedRect(0f,length/21f,1f,-length/21f, 0.015625f,21/64f,0.015625f+((i+1)*length/64f),(length*i)/64f);
